Break settings sections out into separate files
