+++ xcode/Subconscious/Shared/Components/Common/ResourceSyncBadge.swift
@@ -1,0 +1,45 @@
+//
+//  ResourceSyncBadge.swift
+//  Subconscious (iOS)
+//
+//  Created by Ben Follington on 30/6/2023.
+//
+
+import Foundation
+import SwiftUI
+
+struct PendingSyncBadge: View {
+    @State var spin = false
+    
+    var body: some View {
+        Image(systemName: "arrow.triangle.2.circlepath")
+            .rotationEffect(.degrees(spin ? 360 : 0))
+            .animation(Animation.linear
+                .repeatForever(autoreverses: false)
+                .speed(0.4), value: spin)
+            .task{
+                self.spin = true
+            }
+    }
+}
+
+struct ResourceSyncBadge: View {
+    var status: ResourceStatus
+
+    var body: some View {
+        switch status {
+        case .initial:
+            Image(systemName: "arrow.triangle.2.circlepath")
+                .foregroundColor(.secondary)
+        case .pending:
+            PendingSyncBadge()
+                .foregroundColor(.secondary)
+        case .succeeded:
+            Image(systemName: "checkmark.circle")
+                .foregroundColor(.secondary)
+        case .failed:
+            Image(systemName: "exclamationmark.arrow.triangle.2.circlepath")
+                .foregroundColor(.red)
+        }
+    }
+}
